refactor(islamic-website): move constants file to root folder, add & use global contansts, update imports
