+++ apps/islamic-website/app/partnerships/page.tsx
@@ -1,4 +1,4 @@
-import { DEPLOY_URL } from '../../src/contansts';
+import { DEPLOY_URL } from '../../contansts';
 import { Metadata } from 'next';
 
 export { PartnershipPage as default } from '@haqq/islamic-website/partnership-page';
@@ -15,4 +15,3 @@ export const metadata: Metadata = {
   },
   metadataBase: new URL(DEPLOY_URL),
 };
-
