+++ apps/islamic-website/app/wallet/page.tsx
@@ -1,4 +1,4 @@
-import { DEPLOY_URL } from '../../src/contansts';
+import { DEPLOY_URL } from '../../contansts';
 import { Metadata } from 'next';
 
 export { WalletPage as default } from '@haqq/islamic-website/wallet-page';
