--- apps/islamic-website/src/contansts.ts
@@ -1,1 +1,0 @@
-export const DEPLOY_URL = `https://${process.env['NEXT_PUBLIC_VERCEL_URL']}`;
